Separate Type Selector to independent component
